+++ wp-content/themes/shola-jawid/archive-announcement.php
@@ -70,7 +70,7 @@ $paged = max( 1, get_query_var( 'paged' ) );
 					);
 					if ( $links ) {
 						foreach ( $links as $link ) {
-							$link = shola_to_persian_digits( $link );
+							$link = shola_persian_digits_pagination_link( $link );
 							$link = str_replace( 'page-numbers', 'page-num', $link );
 							echo wp_kses_post( $link );
 						}