+++ wp-content/themes/shola-jawid/page-party-documents.php
@@ -73,7 +73,7 @@ $party_documents_query = new WP_Query(
 					);
 					if ( $links ) {
 						foreach ( $links as $link ) {
-							$link = shola_to_persian_digits( $link );
+							$link = shola_persian_digits_pagination_link( $link );
 							$link = str_replace( 'page-numbers', 'page-num', $link );
 							echo wp_kses_post( $link );
 						}